[feat] 자동 로그아웃 임시 팝업 UI 적용 #107
메인 헤더에서 카운트다운 및 팝업창 분리
